GG-20495 Rework GridCommandHandlerTest.
        (cherry picked from commit 932b66bd5f322826f42704278e178c7c3f7ba6f7)
