+++ modules/core/src/main/java/org/apache/ignite/internal/visor/verify/VisorIdleVerifyTaskArg.java
@@ -25,6 +25,9 @@ import org.apache.ignite.internal.util.typedef.internal.S;
 import org.apache.ignite.internal.util.typedef.internal.U;
 import org.apache.ignite.internal.visor.VisorDataTransferObject;
 
+import static java.util.Objects.nonNull;
+import static org.apache.ignite.internal.visor.verify.CacheFilterEnum.DEFAULT;
+
 /**
  * Arguments for task {@link VisorIdleVerifyTask}.
  * <br/>
@@ -46,7 +49,7 @@ public class VisorIdleVerifyTaskArg extends VisorDataTransferObject {
     private boolean skipZeros;
 
     /** Cache kind. */
-    private CacheFilterEnum cacheFilterEnum;
+    private CacheFilterEnum cacheFilterEnum = DEFAULT;
 
     /**
      * Default constructor.
@@ -59,7 +62,7 @@ public class VisorIdleVerifyTaskArg extends VisorDataTransferObject {
      * @param caches Caches.
      * @param excludeCaches Exclude caches or group.
      * @param skipZeros Skip zeros partitions.
-     * @param cacheFilterEnum Cache kind.
+     * @param cacheFilterEnum Cache kind, require non null.
      * @param checkCrc Check CRC sum on stored pages on disk.
      */
     public VisorIdleVerifyTaskArg(
@@ -69,11 +72,13 @@ public class VisorIdleVerifyTaskArg extends VisorDataTransferObject {
         CacheFilterEnum cacheFilterEnum,
         boolean checkCrc
     ) {
+        assert nonNull(cacheFilterEnum) : "Cache filter can't be null";
+
         this.caches = caches;
         this.excludeCaches = excludeCaches;
         this.skipZeros = skipZeros;
-        this.cacheFilterEnum = (cacheFilterEnum == null ? CacheFilterEnum.DEFAULT : cacheFilterEnum);
         this.checkCrc = checkCrc;
+        this.cacheFilterEnum = cacheFilterEnum;
     }
 
     /**
@@ -81,15 +86,14 @@ public class VisorIdleVerifyTaskArg extends VisorDataTransferObject {
      * @param checkCrc Check CRC sum on stored pages on disk.
      */
     public VisorIdleVerifyTaskArg(Set<String> caches, boolean checkCrc) {
-        this.caches = caches;
-        this.checkCrc = checkCrc;
+        this(caches, null, false, DEFAULT, false);
     }
 
     /**
      * @param caches Caches.
      */
     public VisorIdleVerifyTaskArg(Set<String> caches) {
-        this.caches = caches;
+        this(caches, null, false, DEFAULT, false);
     }
 
     /** */
@@ -160,9 +164,7 @@ public class VisorIdleVerifyTaskArg extends VisorDataTransferObject {
             if (protoVer >= V4) {
                 skipZeros = in.readBoolean();
 
-                CacheFilterEnum cfe = CacheFilterEnum.fromOrdinal(in.readByte());
-
-                cacheFilterEnum = (cfe == null ? CacheFilterEnum.DEFAULT : cfe);
+                cacheFilterEnum = CacheFilterEnum.fromOrdinal(in.readByte());
             }
         }
     }
@@ -191,7 +193,9 @@ public class VisorIdleVerifyTaskArg extends VisorDataTransferObject {
 
     /** */
     protected void cacheFilterEnum(CacheFilterEnum cacheFilterEnum) {
-        this.cacheFilterEnum = (cacheFilterEnum == null ? CacheFilterEnum.DEFAULT : cacheFilterEnum);
+        assert nonNull(cacheFilterEnum);
+
+        this.cacheFilterEnum = cacheFilterEnum;
     }
 
     /**
