+++ modules/core/src/test/java/org/apache/ignite/util/GridCommandHandlerSslTest.java
@@ -37,7 +37,7 @@ import static org.apache.ignite.testframework.GridTestUtils.assertContains;
 /**
  * Command line handler test with SSL.
  */
-public class GridCommandHandlerSslTest extends GridCommandHandlerAbstractTest {
+public class GridCommandHandlerSslTest extends GridCommandHandlerClusterPerMethodAbstractTest {
     /** */
     private volatile String[] cipherSuites;
