+++ modules/core/src/test/java/org/apache/ignite/util/GridCommandHandlerTest.java
@@ -17,48 +17,14 @@
 
 package org.apache.ignite.util;
 
-import javax.cache.processor.EntryProcessor;
-import javax.cache.processor.EntryProcessorException;
-import javax.cache.processor.MutableEntry;
-import java.io.File;
-import java.io.IOException;
-import java.io.RandomAccessFile;
-import java.io.Serializable;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.Arrays;
-import java.util.Collection;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Map;
-import java.util.Set;
-import java.util.TreeMap;
-import java.util.UUID;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.ExecutorService;
-import java.util.concurrent.Executors;
-import java.util.concurrent.ThreadLocalRandom;
-import java.util.concurrent.TimeUnit;
-import java.util.concurrent.atomic.AtomicBoolean;
-import java.util.concurrent.atomic.AtomicInteger;
-import java.util.concurrent.atomic.LongAdder;
-import java.util.regex.Matcher;
-import java.util.regex.Pattern;
-import java.util.stream.Collectors;
-import java.util.stream.Stream;
 import org.apache.ignite.Ignite;
 import org.apache.ignite.IgniteAtomicSequence;
 import org.apache.ignite.IgniteCache;
 import org.apache.ignite.IgniteCheckedException;
 import org.apache.ignite.IgniteDataStreamer;
 import org.apache.ignite.IgniteException;
-import org.apache.ignite.cache.CacheMode;
-import org.apache.ignite.cache.PartitionLossPolicy;
 import org.apache.ignite.cache.affinity.rendezvous.RendezvousAffinityFunction;
 import org.apache.ignite.cluster.ClusterNode;
-import org.apache.ignite.configuration.AtomicConfiguration;
 import org.apache.ignite.configuration.CacheConfiguration;
 import org.apache.ignite.configuration.DataRegionConfiguration;
 import org.apache.ignite.internal.GridJobExecuteResponse;
@@ -67,12 +33,8 @@ import org.apache.ignite.internal.IgniteInternalFuture;
 import org.apache.ignite.internal.TestRecordingCommunicationSpi;
 import org.apache.ignite.internal.client.util.GridConcurrentHashSet;
 import org.apache.ignite.internal.commandline.CommandHandler;
-import org.apache.ignite.internal.commandline.CommandList;
-import org.apache.ignite.internal.commandline.argument.CommandArg;
-import org.apache.ignite.internal.commandline.cache.CacheSubcommands;
 import org.apache.ignite.internal.managers.communication.GridIoMessage;
 import org.apache.ignite.internal.pagemem.wal.record.DataEntry;
-import org.apache.ignite.internal.processors.cache.CacheGroupContext;
 import org.apache.ignite.internal.processors.cache.CacheObjectImpl;
 import org.apache.ignite.internal.processors.cache.GridCacheContext;
 import org.apache.ignite.internal.processors.cache.GridCacheEntryEx;
@@ -87,16 +49,13 @@ import org.apache.ignite.internal.processors.cache.persistence.GridCacheDatabase
 import org.apache.ignite.internal.processors.cache.persistence.diagnostic.pagelocktracker.dumpprocessors.ToFileDumpProcessor;
 import org.apache.ignite.internal.processors.cache.transactions.IgniteInternalTx;
 import org.apache.ignite.internal.processors.cache.transactions.IgniteTxEntry;
-import org.apache.ignite.internal.processors.cache.transactions.IgniteTxManager;
 import org.apache.ignite.internal.processors.cache.transactions.TransactionProxyImpl;
 import org.apache.ignite.internal.processors.cache.version.GridCacheVersion;
-import org.apache.ignite.internal.processors.datastructures.GridCacheInternalKeyImpl;
 import org.apache.ignite.internal.util.lang.GridAbsPredicate;
 import org.apache.ignite.internal.util.typedef.F;
 import org.apache.ignite.internal.util.typedef.G;
 import org.apache.ignite.internal.util.typedef.T2;
 import org.apache.ignite.internal.util.typedef.X;
-import org.apache.ignite.internal.util.typedef.internal.CU;
 import org.apache.ignite.internal.util.typedef.internal.U;
 import org.apache.ignite.internal.visor.tx.VisorTxInfo;
 import org.apache.ignite.internal.visor.tx.VisorTxTaskResult;
@@ -108,26 +67,45 @@ import org.apache.ignite.plugin.extensions.communication.Message;
 import org.apache.ignite.testframework.GridTestUtils;
 import org.apache.ignite.transactions.Transaction;
 import org.apache.ignite.transactions.TransactionRollbackException;
-import org.apache.ignite.transactions.TransactionState;
 import org.apache.ignite.transactions.TransactionTimeoutException;
 import org.jetbrains.annotations.NotNull;
 import org.junit.Test;
 
+import javax.cache.processor.EntryProcessor;
+import javax.cache.processor.EntryProcessorException;
+import javax.cache.processor.MutableEntry;
+import java.io.File;
+import java.io.IOException;
+import java.io.RandomAccessFile;
+import java.io.Serializable;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.Collection;
+import java.util.List;
+import java.util.Map;
+import java.util.Set;
+import java.util.TreeMap;
+import java.util.UUID;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.ThreadLocalRandom;
+import java.util.concurrent.atomic.AtomicBoolean;
+import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.LongAdder;
+import java.util.regex.Matcher;
+import java.util.regex.Pattern;
+import java.util.stream.Collectors;
+
 import static java.io.File.separatorChar;
-import static java.util.Arrays.asList;
-import static org.apache.ignite.IgniteSystemProperties.IGNITE_ENABLE_EXPERIMENTAL_COMMAND;
 import static org.apache.ignite.cache.CacheAtomicityMode.TRANSACTIONAL;
+import static org.apache.ignite.cache.CacheMode.PARTITIONED;
 import static org.apache.ignite.cache.CacheWriteSynchronizationMode.FULL_SYNC;
+import static org.apache.ignite.cache.PartitionLossPolicy.READ_ONLY_SAFE;
 import static org.apache.ignite.internal.commandline.CommandHandler.EXIT_CODE_OK;
 import static org.apache.ignite.internal.commandline.CommandHandler.EXIT_CODE_UNEXPECTED_ERROR;
-import static org.apache.ignite.internal.commandline.CommandHandler.UTILITY_NAME;
-import static org.apache.ignite.internal.commandline.CommandList.WAL;
-import static org.apache.ignite.internal.commandline.OutputFormat.MULTI_LINE;
-import static org.apache.ignite.internal.commandline.OutputFormat.SINGLE_LINE;
-import static org.apache.ignite.internal.commandline.cache.CacheSubcommands.HELP;
 import static org.apache.ignite.internal.processors.diagnostic.DiagnosticProcessor.DEFAULT_TARGET_FOLDER;
 import static org.apache.ignite.testframework.GridTestUtils.assertContains;
-import static org.apache.ignite.testframework.GridTestUtils.assertNotContains;
 import static org.apache.ignite.testframework.GridTestUtils.waitForCondition;
 import static org.apache.ignite.transactions.TransactionConcurrency.OPTIMISTIC;
 import static org.apache.ignite.transactions.TransactionConcurrency.PESSIMISTIC;
@@ -135,13 +113,15 @@ import static org.apache.ignite.transactions.TransactionIsolation.READ_COMMITTED
 
 /**
  * Command line handler test.
+ * You can use this class if you need create nodes for each test.
+ * If you not necessary create nodes for each test you can try use {@link GridCommandHandlerClusterByClassTest}
  */
-public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
+public class GridCommandHandlerTest extends GridCommandHandlerClusterPerMethodAbstractTest {
+    /** */
+    protected static File defaultDiagnosticDir;
 
     /** */
-    private File defaultDiagnosticDir;
-    /** */
-    private File customDiagnosticDir;
+    protected static File customDiagnosticDir;
 
     /** {@inheritDoc} */
     @Override protected void beforeTest() throws Exception {
@@ -153,8 +133,8 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
     }
 
     /** {@inheritDoc} */
-    @Override protected void afterTest() throws Exception {
-        super.afterTest();
+    @Override protected void cleanPersistenceDir() throws Exception {
+        super.cleanPersistenceDir();
 
         cleanDiagnosticDir();
     }
@@ -162,7 +142,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
     /**
      * @throws IgniteCheckedException If failed.
      */
-    private void initDiagnosticDir() throws IgniteCheckedException {
+    protected void initDiagnosticDir() throws IgniteCheckedException {
         defaultDiagnosticDir = new File(U.defaultWorkDirectory()
             + separatorChar + DEFAULT_TARGET_FOLDER + separatorChar);
 
@@ -173,20 +153,17 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
     /**
      * Clean diagnostic directories.
      */
-    private void cleanDiagnosticDir() {
+    protected void cleanDiagnosticDir() {
         U.delete(defaultDiagnosticDir);
         U.delete(customDiagnosticDir);
     }
 
-    /** {@inheritDoc} */
-    @Override public String getTestIgniteInstanceName() {
-        return "gridCommandHandlerTest";
-    }
     /**
      * Test activation works via control.sh
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testActivate() throws Exception {
         Ignite ignite = startGrids(1);
 
@@ -202,6 +179,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testDeactivate() throws Exception {
         Ignite ignite = startGrids(1);
 
@@ -221,6 +199,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testState() throws Exception {
         Ignite ignite = startGrids(1);
 
@@ -238,6 +217,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testBaselineCollect() throws Exception {
         Ignite ignite = startGrids(1);
 
@@ -255,6 +235,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testBaselineCollectCrd() throws Exception {
         Ignite ignite = startGrids(2);
 
@@ -273,8 +254,6 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
 
         stopGrid(0);
 
-        testOut.reset();
-
         assertEquals(EXIT_CODE_OK, execute("--baseline", "--port", "11212"));
 
         crdStr = findCrdInfo();
@@ -283,8 +262,6 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
             grid(1).cluster().localNode().consistentId() + ", Order=2)", crdStr);
 
         startGrid(0);
-
-        testOut.reset();
 
         assertEquals(EXIT_CODE_OK, execute("--baseline", "--port", "11212"));
 
@@ -295,50 +272,12 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
 
         stopGrid(1);
 
-        testOut.reset();
-
         assertEquals(EXIT_CODE_OK, execute("--baseline", "--port", "11211"));
 
         crdStr = findCrdInfo();
 
         assertEquals("(Coordinator: ConsistentId=" +
             grid(0).cluster().localNode().consistentId() + ", Order=4)", crdStr);
-
-    }
-
-    /**
-     * Very basic tests for running the command in different enviroment which other command are running in.
-     *
-     * @throws Exception If failed.
-     */
-    public void testFindAndDeleteGarbage() throws Exception {
-        Ignite ignite = startGrids(2);
-
-        ignite.cluster().active(true);
-
-        injectTestSystemOut();
-
-        ignite.createCaches(Arrays.asList(
-            new CacheConfiguration<>("garbage1").setGroupName("groupGarbage"),
-            new CacheConfiguration<>("garbage2").setGroupName("groupGarbage")));
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "find_garbage", "--port", "11212"));
-
-        assertContains(log, testOut.toString(), "garbage not found");
-
-        testOut.reset();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "find_garbage",
-            ignite(0).localNode().id().toString(), "--port", "11212"));
-
-        assertContains(log, testOut.toString(), "garbage not found");
-
-        testOut.reset();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "find_garbage",
-            "groupGarbage", "--port", "11212"));
-
-        assertContains(log, testOut.toString(), "garbage not found");
     }
 
     /**
@@ -380,6 +319,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testBaselineAdd() throws Exception {
         Ignite ignite = startGrids(1);
 
@@ -400,6 +340,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testBaselineRemove() throws Exception {
         Ignite ignite = startGrids(1);
         Ignite other = startGrid("nodeToStop");
@@ -423,6 +364,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testBaselineRemoveOnNotActiveCluster() throws Exception {
         Ignite ignite = startGrids(1);
         Ignite other = startGrid("nodeToStop");
@@ -455,6 +397,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testBaselineSet() throws Exception {
         Ignite ignite = startGrids(1);
 
@@ -476,6 +419,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testBaselineVersion() throws Exception {
         Ignite ignite = startGrids(1);
 
@@ -497,6 +441,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testActiveTransactions() throws Exception {
         Ignite ignite = startGridsMultiThreaded(2);
 
@@ -652,205 +597,9 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
     }
 
     /**
-     * Smoke test for --tx --info command.
-     */
-    public void testTransactionInfo() throws Exception {
-        Ignite ignite = startGridsMultiThreaded(5);
-
-        ignite.cluster().active(true);
-
-        Ignite client = startGrid("client");
-
-        client.getOrCreateCache(new CacheConfiguration<>(DEFAULT_CACHE_NAME)
-            .setAtomicityMode(TRANSACTIONAL).setBackups(2).setWriteSynchronizationMode(FULL_SYNC));
-
-        for (Ignite ig : G.allGrids())
-            assertNotNull(ig.cache(DEFAULT_CACHE_NAME));
-
-        CountDownLatch lockLatch = new CountDownLatch(1);
-        CountDownLatch unlockLatch = new CountDownLatch(1);
-
-        IgniteInternalFuture<?> fut = startTransactions("testTransactionInfo", lockLatch, unlockLatch, false);
-
-        try {
-            U.awaitQuiet(lockLatch);
-
-            doSleep(3000); // Should be more than enough for all transactions to appear in contexts.
-
-            Set<GridCacheVersion> nearXids = new HashSet<>();
-
-            for (int i = 0; i < 5; i++) {
-                IgniteEx grid = grid(i);
-
-                IgniteTxManager tm = grid.context().cache().context().tm();
-
-                for (IgniteInternalTx tx : tm.activeTransactions())
-                    nearXids.add(tx.nearXidVersion());
-            }
-
-            injectTestSystemOut();
-
-            for (GridCacheVersion nearXid : nearXids)
-                assertEquals(EXIT_CODE_OK, execute("--tx", "--info", nearXid.toString()));
-
-            String out = testOut.toString();
-
-            for (GridCacheVersion nearXid : nearXids)
-                assertTrue(out.contains(nearXid.toString()));
-        }
-        finally {
-            unlockLatch.countDown();
-
-            fut.get();
-        }
-    }
-
-    /**
-     * Smoke test for historical mode of --tx --info command.
-     */
-    public void testTransactionHistoryInfo() throws Exception {
-        Ignite ignite = startGridsMultiThreaded(2);
-
-        ignite.cluster().active(true);
-
-        Ignite client = startGrid("client");
-
-        client.getOrCreateCache(new CacheConfiguration<>(DEFAULT_CACHE_NAME)
-            .setAtomicityMode(TRANSACTIONAL).setBackups(2).setWriteSynchronizationMode(FULL_SYNC));
-
-        for (Ignite ig : G.allGrids())
-            assertNotNull(ig.cache(DEFAULT_CACHE_NAME));
-
-        CountDownLatch lockLatch = new CountDownLatch(1);
-        CountDownLatch unlockLatch = new CountDownLatch(1);
-
-        IgniteInternalFuture<?> fut = startTransactions("testTransactionHistoryInfo", lockLatch, unlockLatch, false);
-
-        U.awaitQuiet(lockLatch);
-
-        doSleep(3000); // Should be more than enough for all transactions to appear in contexts.
-
-        Set<GridCacheVersion> nearXids = new HashSet<>();
-
-        for (int i = 0; i < 2; i++) {
-            IgniteEx grid = grid(i);
-
-            IgniteTxManager tm = grid.context().cache().context().tm();
-
-            for (IgniteInternalTx tx : tm.activeTransactions())
-                nearXids.add(tx.nearXidVersion());
-        }
-
-        unlockLatch.countDown();
-
-        fut.get();
-
-        doSleep(3000); // Should be more than enough for all transactions to disappear from contexts after finish.
-
-        injectTestSystemOut();
-
-        boolean commitMatched = false;
-        boolean rollbackMatched = false;
-
-        for (GridCacheVersion nearXid : nearXids) {
-            assertEquals(EXIT_CODE_OK, execute("--tx", "--info", nearXid.toString()));
-
-            String out = testOut.toString();
-
-            testOut.reset();
-
-            assertContains(log, out, "Transaction was found in completed versions history of the following nodes:");
-
-            if (out.contains(TransactionState.COMMITTED.name())) {
-                commitMatched = true;
-
-                assertNotContains(log, out, TransactionState.ROLLED_BACK.name());
-            }
-
-            if (out.contains(TransactionState.ROLLED_BACK.name())) {
-                rollbackMatched = true;
-
-                assertNotContains(log, out, TransactionState.COMMITTED.name());
-            }
-
-        }
-
-        assertTrue(commitMatched);
-        assertTrue(rollbackMatched);
-    }
-
-    /**
-     *
-     */
-    public void testKillHangingLocalTransactions() throws Exception {
-        Ignite ignite = startGridsMultiThreaded(2);
-
-        ignite.cluster().active(true);
-
-        Ignite client = startGrid("client");
-
-        client.getOrCreateCache(new CacheConfiguration<>(DEFAULT_CACHE_NAME).
-            setAtomicityMode(TRANSACTIONAL).
-            setWriteSynchronizationMode(FULL_SYNC).
-            setAffinity(new RendezvousAffinityFunction(false, 64)));
-
-        Ignite prim = primaryNode(0L, DEFAULT_CACHE_NAME);
-
-        // Blocks lock response to near node.
-        TestRecordingCommunicationSpi.spi(prim).blockMessages(GridNearLockResponse.class, client.name());
-
-        TestRecordingCommunicationSpi.spi(client).blockMessages(GridNearTxFinishRequest.class, prim.name());
-
-        GridNearTxLocal clientTx = null;
-
-        try (Transaction tx = client.transactions().txStart(PESSIMISTIC, READ_COMMITTED, 2000, 1)) {
-            clientTx = ((TransactionProxyImpl)tx).tx();
-
-            client.cache(DEFAULT_CACHE_NAME).put(0L, 0L);
-
-            fail();
-        }
-        catch (Exception e) {
-            assertTrue(X.hasCause(e, TransactionTimeoutException.class));
-        }
-
-        assertNotNull(clientTx);
-
-        IgniteEx primEx = (IgniteEx)prim;
-
-        IgniteInternalTx tx0 = primEx.context().cache().context().tm().activeTransactions().iterator().next();
-
-        assertNotNull(tx0);
-
-        CommandHandler h = new CommandHandler();
-
-        validate(h, map -> {
-            ClusterNode node = grid(0).cluster().localNode();
-
-            VisorTxTaskResult res = map.get(node);
-
-            for (VisorTxInfo info : res.getInfos())
-                assertEquals(tx0.xid(), info.getXid());
-
-            assertEquals(1, map.size());
-        }, "--tx", "--xid", tx0.xid().toString(), "--kill");
-
-        tx0.finishFuture().get();
-
-        TestRecordingCommunicationSpi.spi(prim).stopBlock();
-
-        TestRecordingCommunicationSpi.spi(client).stopBlock();
-
-        IgniteInternalFuture<?> nearFinFut = U.field(clientTx, "finishFut");
-
-        nearFinFut.get();
-
-        checkUserFutures();
-    }
-
-    /**
      * Simulate uncommitted backup transactions and test rolling back using utility.
      */
+    @Test
     public void testKillHangingRemoteTransactions() throws Exception {
         final int cnt = 3;
 
@@ -1037,6 +786,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testBaselineAddOnNotActiveCluster() throws Exception {
         Ignite ignite = startGrid(1);
 
@@ -1068,416 +818,8 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
         assertFalse(testOutStr, isInstanse1Found);
     }
 
-    /**
-     */
-    public void testCacheHelp() {
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "help"));
-
-        String output = testOut.toString();
-
-        for (CacheSubcommands cmd : CacheSubcommands.values()) {
-            if (cmd != HELP) {
-                assertContains(log, output, cmd.toString());
-
-                Class<? extends Enum<? extends CommandArg>> args = cmd.getCommandArgs();
-
-                if (args != null)
-                    for (Enum<? extends CommandArg> arg : args.getEnumConstants())
-                        assertTrue(cmd + " " + arg, output.contains(arg.toString()));
-
-            }
-            else {
-                assertContains(log, output, CommandHandler.UTILITY_NAME);
-
-                assertNotContains(log, output, "control.sh");
-            }
-        }
-    }
-
-    /**
-     */
-    public void testCorrectCacheOptionsNaming() {
-        Pattern p = Pattern.compile("^--([a-z]+(-)?)+([a-z]+)");
-
-        for (CacheSubcommands cmd : CacheSubcommands.values()) {
-            Class<? extends Enum<? extends CommandArg>> args = cmd.getCommandArgs();
-
-            if (args != null)
-                for (Enum<? extends CommandArg> arg : args.getEnumConstants())
-                    assertTrue(arg.toString(), p.matcher(arg.toString()).matches());
-        }
-    }
-
-    /**
-     */
-    public void testHelp() {
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--help"));
-
-        String testOutStr = testOut.toString();
-
-        for (CommandList cmd : CommandList.values())
-            assertContains(log, testOutStr, cmd.toString());
-
-        assertContains(log, testOutStr, "Control utility script");
-        assertContains(log, testOutStr, CommandHandler.UTILITY_NAME);
-
-        assertNotContains(log, testOutStr, "Control.sh");
-    }
-
     /** */
     @Test
-    public void testPrintTimestampAtEndsOfExecution() {
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute());
-
-        String testOutStr = testOut.toString();
-
-        assertContains(log, testOutStr, "Control utility has completed execution at: ");
-    }
-
-    /**
-     * @throws Exception If failed.
-     */
-    public void testCacheIdleVerify() throws Exception {
-        IgniteEx ignite = startGrids(2);
-
-        ignite.cluster().active(true);
-
-        createCacheAndPreload(ignite, 100);
-
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify"));
-
-        assertContains(log, testOut.toString(), "no conflicts have been found");
-
-        HashSet<Integer> clearKeys = new HashSet<>(asList(1, 2, 3, 4, 5, 6));
-
-        ignite.context().cache().cache(DEFAULT_CACHE_NAME).clearLocallyAll(clearKeys, true, true, true);
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify"));
-
-        assertContains(log, testOut.toString(), "conflict partitions");
-    }
-
-    /**
-     * @throws Exception If failed.
-     */
-    @Test
-    public void testCacheIdleVerifyPrintLostPartitions() throws Exception {
-        IgniteEx ignite = startGrids(3);
-
-        ignite.cluster().active(true);
-
-        ignite.createCache(new CacheConfiguration<>(DEFAULT_CACHE_NAME)
-            .setAffinity(new RendezvousAffinityFunction(false, 16))
-            .setCacheMode(CacheMode.PARTITIONED)
-            .setPartitionLossPolicy(PartitionLossPolicy.READ_ONLY_SAFE)
-            .setBackups(1));
-
-        try (IgniteDataStreamer streamer = ignite.dataStreamer(DEFAULT_CACHE_NAME)) {
-            for (int i = 0; i < 1000; i++)
-                streamer.addData(i, i);
-        }
-
-        String g1Name = grid(1).name();
-
-        stopGrid(1);
-
-        cleanPersistenceDir(g1Name);
-
-        //Start node 2 with empty PDS. Rebalance will be started.
-        startGrid(1);
-
-        //During rebalance stop node 3. Rebalance will be stopped which lead to lost partitions.
-        stopGrid(2);
-
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify", "--yes"));
-
-        //Lost partitions unexpectedly doesn't detected here.
-        //assertContains(log, testOut.toString(), "LOST partitions:");
-    }
-
-    /**
-     * @throws Exception If failed.
-     */
-    public void testCacheIdleVerifyNodeFilter() throws Exception {
-        IgniteEx ignite = startGrids(3);
-
-        ignite.cluster().active(true);
-
-        Object lastNodeCId = ignite.localNode().consistentId();
-
-        ignite.createCache(new CacheConfiguration<>(DEFAULT_CACHE_NAME)
-            .setAffinity(new RendezvousAffinityFunction(false, 32))
-            .setNodeFilter(node -> !lastNodeCId.equals(node.consistentId()))
-            .setBackups(1));
-
-        try (IgniteDataStreamer streamer = ignite.dataStreamer(DEFAULT_CACHE_NAME)) {
-            for (int i = 0; i < 100; i++)
-                streamer.addData(i, i);
-        }
-
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify", DEFAULT_CACHE_NAME));
-
-        assertContains(log, testOut.toString(), "no conflicts have been found");
-    }
-
-    /**
-     * Tests that both update counter and hash conflicts are detected.
-     *
-     * @throws Exception If failed.
-     */
-    public void testCacheIdleVerifyTwoConflictTypes() throws Exception {
-        IgniteEx ignite = startGrids(2);
-
-        ignite.cluster().active(true);
-
-        createCacheAndPreload(ignite, 100);
-
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify"));
-
-        assertContains(log, testOut.toString(), "no conflicts have been found");
-
-        GridCacheContext<Object, Object> cacheCtx = ignite.cachex(DEFAULT_CACHE_NAME).context();
-
-        corruptDataEntry(cacheCtx, 1, true, false);
-
-        corruptDataEntry(cacheCtx, 1 + cacheCtx.config().getAffinity().partitions() / 2, false, true);
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify"));
-
-        assertContains(log, testOut.toString(), "found 2 conflict partitions");
-    }
-
-    /**
-     * Tests that empty partitions with non-zero update counter are not included into the idle_verify dump.
-     *
-     * @throws Exception If failed.
-     */
-    public void testCacheIdleVerifyDumpSkipZerosUpdateCounters() throws Exception {
-        IgniteEx ignite = startGrids(2);
-
-        ignite.cluster().active(true);
-
-        int emptyPartId = 31;
-
-        // Less than parts number for ability to check skipZeros flag.
-        createCacheAndPreload(ignite, emptyPartId);
-
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify", "--dump", "--skip-zeros", DEFAULT_CACHE_NAME));
-
-        Matcher fileNameMatcher = dumpFileNameMatcher();
-
-        assertTrue(fileNameMatcher.find());
-
-        String zeroUpdateCntrs = new String(Files.readAllBytes(Paths.get(fileNameMatcher.group(1))));
-
-        assertContains(log, zeroUpdateCntrs, "idle_verify check has finished, found " + emptyPartId + " partitions");
-        assertContains(log, zeroUpdateCntrs, "1 partitions was skipped");
-        assertContains(log, zeroUpdateCntrs, "idle_verify check has finished, no conflicts have been found.");
-
-        assertSort(emptyPartId, zeroUpdateCntrs);
-
-        testOut.reset();
-
-        // The result of the following cache operations is that
-        // the size of the 32-th partition is equal to zero and update counter is equal to 2.
-        ignite.cache(DEFAULT_CACHE_NAME).put(emptyPartId, emptyPartId);
-
-        ignite.cache(DEFAULT_CACHE_NAME).remove(emptyPartId, emptyPartId);
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify", "--dump", "--skip-zeros", DEFAULT_CACHE_NAME));
-
-        fileNameMatcher = dumpFileNameMatcher();
-
-        assertTrue(fileNameMatcher.find());
-
-        String nonZeroUpdateCntrs = new String(Files.readAllBytes(Paths.get(fileNameMatcher.group(1))));
-
-        assertContains(log, nonZeroUpdateCntrs, "idle_verify check has finished, found " + 31 + " partitions");
-        assertContains(log, nonZeroUpdateCntrs, "1 partitions was skipped");
-        assertContains(log, nonZeroUpdateCntrs, "idle_verify check has finished, no conflicts have been found.");
-
-        assertSort(31, zeroUpdateCntrs);
-
-        assertEquals(zeroUpdateCntrs, nonZeroUpdateCntrs);
-    }
-
-    /**
-     * Tests that idle verify print partitions info.
-     *
-     * @throws Exception If failed.
-     */
-    public void testCacheIdleVerifyDump() throws Exception {
-        IgniteEx ignite = startGrids(3);
-
-        ignite.cluster().active(true);
-
-        int keysCount = 20;//less than parts number for ability to check skipZeros flag.
-
-        createCacheAndPreload(ignite, keysCount);
-
-        int parts = ignite.affinity(DEFAULT_CACHE_NAME).partitions();
-
-        ignite.createCache(new CacheConfiguration<>()
-            .setAffinity(new RendezvousAffinityFunction(false, parts))
-            .setBackups(1)
-            .setName(DEFAULT_CACHE_NAME + "other"));
-
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify", "--dump", DEFAULT_CACHE_NAME));
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify", "--dump", "--skip-zeros", DEFAULT_CACHE_NAME));
-
-        Matcher fileNameMatcher = dumpFileNameMatcher();
-
-        if (fileNameMatcher.find()) {
-            String dumpWithZeros = new String(Files.readAllBytes(Paths.get(fileNameMatcher.group(1))));
-
-            assertContains(log, dumpWithZeros, "idle_verify check has finished, found " + parts + " partitions");
-            assertContains(log, dumpWithZeros, "Partition: PartitionKeyV2 [grpId=1544803905, grpName=default, partId=0]");
-            assertContains(log, dumpWithZeros, "updateCntr=0, partitionState=OWNING, size=0, partHash=0");
-            assertContains(log, dumpWithZeros, "no conflicts have been found");
-
-            assertSort(parts, dumpWithZeros);
-        }
-
-        if (fileNameMatcher.find()) {
-            String dumpWithoutZeros = new String(Files.readAllBytes(Paths.get(fileNameMatcher.group(1))));
-
-            assertContains(log, dumpWithoutZeros, "idle_verify check has finished, found " + keysCount + " partitions");
-            assertContains(log, dumpWithoutZeros, (parts - keysCount) + " partitions was skipped");
-            assertContains(log, dumpWithoutZeros, "Partition: PartitionKeyV2 [grpId=1544803905, grpName=default, partId=");
-
-            assertNotContains(log, dumpWithoutZeros, "updateCntr=0, partitionState=OWNING, size=0, partHash=0");
-
-            assertContains(log, dumpWithoutZeros, "no conflicts have been found");
-
-            assertSort(keysCount, dumpWithoutZeros);
-        }
-        else
-            fail("Should be found both files");
-    }
-
-    /**
-     * Common method for idle_verify tests with multiple options.
-     *
-     * @throws Exception if failed
-     */
-    public void testCacheIdleVerifyMultipleCacheFilterOptions()
-            throws Exception {
-        IgniteEx ignite = startGrids(2);
-
-        ignite.cluster().active(true);
-
-        ignite.createCache(new CacheConfiguration<>()
-                .setAffinity(new RendezvousAffinityFunction(false, 32))
-                .setGroupName("shared_grp")
-                .setBackups(1)
-                .setName(DEFAULT_CACHE_NAME));
-
-        ignite.createCache(new CacheConfiguration<>()
-                .setAffinity(new RendezvousAffinityFunction(false, 32))
-                .setGroupName("shared_grp")
-                .setBackups(1)
-                .setName(DEFAULT_CACHE_NAME + "_second"));
-
-        ignite.createCache(new CacheConfiguration<>()
-                .setAffinity(new RendezvousAffinityFunction(false, 64))
-                .setBackups(1)
-                .setName(DEFAULT_CACHE_NAME + "_third"));
-
-        ignite.createCache(new CacheConfiguration<>()
-                .setAffinity(new RendezvousAffinityFunction(false, 128))
-                .setBackups(1)
-                .setName("wrong_cache"));
-
-        injectTestSystemOut();
-
-        testCacheIdleVerifyMultipleCacheFilterOptionsCommon(
-            true,
-            "idle_verify check has finished, found",
-            "idle_verify task was executed with the following args: caches=[], excluded=[wrong.*], cacheFilter=[SYSTEM]",
-            "--cache", "idle_verify", "--dump", "--cache-filter", "SYSTEM", "--exclude-caches", "wrong.*"
-        );
-        testCacheIdleVerifyMultipleCacheFilterOptionsCommon(
-            true,
-            "idle_verify check has finished, found 96 partitions",
-            null,
-            "--cache", "idle_verify", "--dump", "--exclude-caches", "wrong.*"
-        );
-        testCacheIdleVerifyMultipleCacheFilterOptionsCommon(
-            true,
-            "idle_verify check has finished, found 32 partitions",
-            null,
-            "--cache", "idle_verify", "--dump", "shared.*"
-        );
-        testCacheIdleVerifyMultipleCacheFilterOptionsCommon(
-            true,
-            "idle_verify check has finished, found 160 partitions",
-            null,
-            "--cache", "idle_verify", "--dump", "shared.*,wrong.*"
-        );
-        testCacheIdleVerifyMultipleCacheFilterOptionsCommon(
-            true,
-            "idle_verify check has finished, found 160 partitions",
-            null,
-            "--cache", "idle_verify", "--dump", "shared.*,wrong.*", "--cache-filter", "USER"
-        );
-        testCacheIdleVerifyMultipleCacheFilterOptionsCommon(
-            true,
-            "idle_verify check has finished, found 160 partitions",
-            null,
-            "--cache", "idle_verify", "--dump", "shared.*,wrong.*"
-        );
-        testCacheIdleVerifyMultipleCacheFilterOptionsCommon(
-            true,
-            "There are no caches matching given filter options",
-            null,
-            "--cache", "idle_verify", "--exclude-caches", ".*"
-        );
-        testCacheIdleVerifyMultipleCacheFilterOptionsCommon(
-            false,
-            "Invalid cache name regexp",
-            null,
-            "--cache", "idle_verify", "--dump", "--exclude-caches", "["
-        );
-        testCacheIdleVerifyMultipleCacheFilterOptionsCommon(
-            true,
-            "idle_verify check has finished, no conflicts have been found.",
-            null,
-            "--cache", "idle_verify", "--exclude-caches", "wrong.*"
-        );
-        testCacheIdleVerifyMultipleCacheFilterOptionsCommon(
-            true,
-            "idle_verify check has finished, no conflicts have been found.",
-            null,
-            "--cache", "idle_verify", "--dump", "--cache-filter", "PERSISTENT"
-        );
-        testCacheIdleVerifyMultipleCacheFilterOptionsCommon(
-            true,
-            "There are no caches matching given filter options.",
-            null,
-            "--cache", "idle_verify", "--cache-filter", "NOT_PERSISTENT"
-        );
-    }
-
-    /** */
     public void testIdleVerifyCheckCrcFailsOnNotIdleCluster() throws Exception {
         checkpointFreq = 100L;
 
@@ -1532,154 +874,11 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
     }
 
     /**
-     * Runs idle_verify with specified arguments and checks the dump if dump option was present.
-     *
-     * @param exitOk whether CommandHandler should exit without errors
-     * @param outputExp expected dump output
-     * @param cmdExp expected command built from command line arguments
-     * @param args command handler arguments
-     * @throws IOException if some of file operations failed
-     */
-    private void testCacheIdleVerifyMultipleCacheFilterOptionsCommon(
-        boolean exitOk,
-        String outputExp,
-        String cmdExp,
-        String... args
-    ) throws IOException {
-        testOut.reset();
-
-        Set<String> argsSet = new HashSet<>(asList(args));
-
-        int exitCode = execute(args);
-
-        assertEquals(testOut.toString(), exitOk, EXIT_CODE_OK == exitCode);
-
-        if (exitCode == EXIT_CODE_OK) {
-            Matcher fileNameMatcher = dumpFileNameMatcher();
-
-            if (fileNameMatcher.find()) {
-                assertContains(log, argsSet, "--dump");
-
-                Path filePath = Paths.get(fileNameMatcher.group(1));
-
-                String dump = new String(Files.readAllBytes(filePath));
-
-                Files.delete(filePath);
-
-                assertContains(log, dump, outputExp);
-
-                if (cmdExp != null)
-                    assertContains(log, dump, cmdExp);
-            }
-            else {
-                assertNotContains(log, argsSet, "--dump");
-
-                assertContains(log, testOut.toString(), outputExp);
-            }
-        } else
-            assertContains(log, testOut.toString(), outputExp);
-    }
-
-    /**
-     * Checking sorting of partitions.
-     *
-     * @param expectedPartsCount Expected parts count.
-     * @param output Output.
-     */
-    private void assertSort(int expectedPartsCount, String output) {
-        Pattern partIdPattern = Pattern.compile(".*partId=([0-9]*)");
-        Pattern primaryPattern = Pattern.compile("Partition instances: \\[PartitionHashRecordV2 \\[isPrimary=true");
-
-        Matcher partIdMatcher = partIdPattern.matcher(output);
-        Matcher primaryMatcher = primaryPattern.matcher(output);
-
-        int i = 0;
-
-        while (partIdMatcher.find()) {
-            assertEquals(i++, Integer.parseInt(partIdMatcher.group(1)));
-            assertTrue(primaryMatcher.find());//primary node should be first in every line
-        }
-
-        assertEquals(expectedPartsCount, i);
-    }
-
-    /**
-     * Tests that idle verify print partitions info.
-     *
-     * @throws Exception If failed.
-     */
-    public void testCacheIdleVerifyDumpForCorruptedData() throws Exception {
-        IgniteEx ignite = startGrids(3);
-
-        ignite.cluster().active(true);
-
-        createCacheAndPreload(ignite, 100);
-
-        injectTestSystemOut();
-
-        corruptingAndCheckDefaultCache(ignite, "USER", true);
-    }
-
-    /**
-     * Tests that idle verify print partitions info while launched without dump option.
-     *
-     * @throws Exception If failed.
-     */
-    public void testCacheIdleVerifyForCorruptedData() throws Exception {
-        IgniteEx ignite = startGrids(3);
-
-        ignite.cluster().active(true);
-
-        createCacheAndPreload(ignite, 100);
-
-        injectTestSystemOut();
-
-        corruptingAndCheckDefaultCache(ignite, "USER", false);
-    }
-
-    /**
-     * @param ignite Ignite.
-     * @param cacheFilter cacheFilter.
-     * @param dump Whether idle_verify should be launched with dump option or not.
-     */
-    private void corruptingAndCheckDefaultCache(IgniteEx ignite, String cacheFilter, boolean dump) throws IOException {
-        injectTestSystemOut();
-
-        GridCacheContext<Object, Object> cacheCtx = ignite.cachex(DEFAULT_CACHE_NAME).context();
-
-        corruptDataEntry(cacheCtx, 0, true, false);
-
-        corruptDataEntry(cacheCtx, cacheCtx.config().getAffinity().partitions() / 2, false, true);
-
-        String resReport = null;
-
-        if (dump) {
-            assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify", "--dump", "--cache-filter", cacheFilter));
-
-            Matcher fileNameMatcher = dumpFileNameMatcher();
-
-            if (fileNameMatcher.find()) {
-                resReport = new String(Files.readAllBytes(Paths.get(fileNameMatcher.group(1))));
-
-                log.info(resReport);
-            }
-            else
-                fail("Should be found dump with conflicts");
-        }
-        else {
-            assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify", "--cache-filter", cacheFilter));
-
-            resReport = testOut.toString();
-        }
-
-        assertContains(log, resReport, "found 2 conflict partitions: [counterConflicts=1, hashConflicts=1]");
-    }
-
-    /**
      * Tests that idle verify print partitions info when node failing.
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testCacheIdleVerifyDumpWhenNodeFailing() throws Exception {
         Ignite ignite = startGrids(3);
 
@@ -1715,6 +914,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testCacheIdleVerifyDumpWhenSeveralNodesFailing() throws Exception {
         int nodes = 6;
 
@@ -1761,6 +961,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
     }
 
     /** */
+    @Test
     public void testCacheIdleVerifyCrcWithCorruptedPartition() throws Exception {
         testCacheIdleVerifyWithCorruptedPartition("--cache", "idle_verify", "--check-crc");
 
@@ -1771,6 +972,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
     }
 
     /** */
+    @Test
     public void testCacheIdleVerifyDumpCrcWithCorruptedPartition() throws Exception {
         testCacheIdleVerifyWithCorruptedPartition("--cache", "idle_verify", "--dump", "--check-crc");
 
@@ -1781,12 +983,12 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
         String dumpFile = parts[1].split("\\.")[0] + ".txt";
 
         for (String line : Files.readAllLines(new File(dumpFile).toPath()))
-            System.out.println(line);
+            testOut.write(line.getBytes());
 
         String outputStr = testOut.toString();
 
-        assertTrue(outputStr, outputStr.contains("idle_verify failed on 1 node."));
-        assertTrue(outputStr, outputStr.contains("idle_verify check has finished, no conflicts have been found."));
+        assertContains(log,outputStr, "idle_verify failed on 1 node.");
+        assertContains(log, outputStr, "idle_verify check has finished, no conflicts have been found.");
     }
 
     /** */
@@ -1838,23 +1040,6 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
     }
 
     /**
-     * Creates default cache and preload some data entries.
-     *
-     * @param ignite Ignite.
-     * @param countEntries Count of entries.
-     */
-    private void createCacheAndPreload(Ignite ignite, int countEntries) {
-        ignite.createCache(new CacheConfiguration<>(DEFAULT_CACHE_NAME)
-            .setAffinity(new RendezvousAffinityFunction(false, 32))
-            .setBackups(1));
-
-        try (IgniteDataStreamer streamer = ignite.dataStreamer(DEFAULT_CACHE_NAME)) {
-            for (int i = 0; i < countEntries; i++)
-                streamer.addData(i, i);
-        }
-    }
-
-    /**
      * Try to finds node failed exception message on output report.
      *
      * @param unstableNodeId Unstable node id.
@@ -1874,88 +1059,11 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
     }
 
     /**
-     * Tests that idle verify print partitions info over system caches.
-     *
-     * @throws Exception If failed.
-     */
-    public void testCacheIdleVerifyDumpForCorruptedDataOnSystemCache() throws Exception {
-        int parts = 32;
-
-        atomicConfiguration = new AtomicConfiguration()
-            .setAffinity(new RendezvousAffinityFunction(false, parts))
-            .setBackups(2);
-
-        IgniteEx ignite = startGrids(3);
-
-        ignite.cluster().active(true);
-
-        if (!isSystemOutAlreadyInjected())
-            injectTestSystemOut();
-
-        // Adding some assignments without deployments.
-        for (int i = 0; i < 100; i++) {
-            ignite.semaphore("s" + i, i, false, true);
-
-            ignite.atomicSequence("sq" + i, 0, true)
-                .incrementAndGet();
-        }
-
-        CacheGroupContext storedSysCacheCtx = ignite.context().cache().cacheGroup(CU.cacheId("default-ds-group"));
-
-        assertNotNull(storedSysCacheCtx);
-
-        corruptDataEntry(storedSysCacheCtx.caches().get(0), new GridCacheInternalKeyImpl("sq0",
-            "default-ds-group"), true, false);
-
-        corruptDataEntry(storedSysCacheCtx.caches().get(0), new GridCacheInternalKeyImpl("sq" + parts / 2,
-            "default-ds-group"), false, true);
-
-        CacheGroupContext memorySysCacheCtx = ignite.context().cache().cacheGroup(CU.cacheId("default-volatile-ds-group"));
-
-        assertNotNull(memorySysCacheCtx);
-
-        corruptDataEntry(memorySysCacheCtx.caches().get(0), new GridCacheInternalKeyImpl("s0",
-            "default-volatile-ds-group"), true, false);
-
-        corruptDataEntry(memorySysCacheCtx.caches().get(0), new GridCacheInternalKeyImpl("s" + parts / 2,
-            "default-volatile-ds-group"), false, true);
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify", "--dump", "--cache-filter", "SYSTEM"));
-
-        Matcher fileNameMatcher = dumpFileNameMatcher();
-
-        if (fileNameMatcher.find()) {
-            String dumpWithConflicts = new String(Files.readAllBytes(Paths.get(fileNameMatcher.group(1))));
-
-            U.log(log, dumpWithConflicts);
-
-            assertContains(log, dumpWithConflicts, "found 4 conflict partitions: [counterConflicts=2, " +
-                "hashConflicts=2]");
-        }
-        else
-            fail("Should be found dump with conflicts");
-    }
-
-    /**
-     * Tests that idle verify print partitions info over persistence client caches.
-     *
-     * @throws Exception If failed.
-     */
-    public void testCacheIdleVerifyDumpForCorruptedDataOnPersistenceClientCache() throws Exception {
-        IgniteEx ignite = startGrids(3);
-
-        ignite.cluster().active(true);
-
-        createCacheAndPreload(ignite, 100);
-
-        corruptingAndCheckDefaultCache(ignite, "PERSISTENT", true);
-    }
-
-    /**
      * Tests that idle verify print partitions info over none-persistence client caches.
      *
      * @throws Exception If failed.
      */
+    @Test
     public void testCacheIdleVerifyDumpForCorruptedDataOnNonePersistenceClientCache() throws Exception {
         int parts = 32;
 
@@ -2002,92 +1110,6 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
     }
 
     /**
-     * Tests that idle verify print partitions info with exclude cache group.
-     *
-     * @throws Exception If failed.
-     */
-    public void testCacheIdleVerifyDumpExcludedCacheGrp() throws Exception {
-        IgniteEx ignite = startGrids(3);
-
-        ignite.cluster().active(true);
-
-        int parts = 32;
-
-        IgniteCache<Object, Object> cache = ignite.createCache(new CacheConfiguration<>()
-            .setAffinity(new RendezvousAffinityFunction(false, parts))
-            .setGroupName("shared_grp")
-            .setBackups(1)
-            .setName(DEFAULT_CACHE_NAME));
-
-        IgniteCache<Object, Object> secondCache = ignite.createCache(new CacheConfiguration<>()
-            .setAffinity(new RendezvousAffinityFunction(false, parts))
-            .setGroupName("shared_grp")
-            .setBackups(1)
-            .setName(DEFAULT_CACHE_NAME + "_second"));
-
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify", "--dump", "--exclude-caches", "shared_grp"));
-
-        Matcher fileNameMatcher = dumpFileNameMatcher();
-
-        if (fileNameMatcher.find()) {
-            String dumpWithConflicts = new String(Files.readAllBytes(Paths.get(fileNameMatcher.group(1))));
-
-            assertContains(log, dumpWithConflicts, "There are no caches matching given filter options");
-        }
-        else
-            fail("Should be found dump with conflicts");
-    }
-
-    /**
-     * Tests that idle verify print partitions info with exclude caches.
-     *
-     * @throws Exception If failed.
-     */
-    public void testCacheIdleVerifyDumpExcludedCaches() throws Exception {
-        IgniteEx ignite = startGrids(3);
-
-        ignite.cluster().active(true);
-
-        int parts = 32;
-
-        ignite.createCache(new CacheConfiguration<>()
-            .setAffinity(new RendezvousAffinityFunction(false, parts))
-            .setGroupName("shared_grp")
-            .setBackups(1)
-            .setName(DEFAULT_CACHE_NAME));
-
-        ignite.createCache(new CacheConfiguration<>()
-            .setAffinity(new RendezvousAffinityFunction(false, parts))
-            .setGroupName("shared_grp")
-            .setBackups(1)
-            .setName(DEFAULT_CACHE_NAME + "_second"));
-
-        ignite.createCache(new CacheConfiguration<>()
-            .setAffinity(new RendezvousAffinityFunction(false, parts))
-            .setBackups(1)
-            .setName(DEFAULT_CACHE_NAME + "_third"));
-
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify", "--dump", "--exclude-caches", DEFAULT_CACHE_NAME
-            + "," + DEFAULT_CACHE_NAME + "_second"));
-
-        Matcher fileNameMatcher = dumpFileNameMatcher();
-
-        if (fileNameMatcher.find()) {
-            String dumpWithConflicts = new String(Files.readAllBytes(Paths.get(fileNameMatcher.group(1))));
-
-            assertContains(log, dumpWithConflicts, "idle_verify check has finished, found 32 partitions");
-            assertContains(log, dumpWithConflicts, "default_third");
-            assertNotContains(log, dumpWithConflicts, "shared_grp");
-        }
-        else
-            fail("Should be found dump with conflicts");
-    }
-
-    /**
      * @return Build matcher for dump file name.
      */
     @NotNull private Matcher dumpFileNameMatcher() {
@@ -2099,6 +1121,7 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
     /**
      * @throws Exception If failed.
      */
+    @Test
     public void testCacheIdleVerifyMovingParts() throws Exception {
         IgniteEx ignite = startGrids(2);
 
@@ -2130,84 +1153,8 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
         assertContains(log, testOut.toString(), "MOVING partitions");
     }
 
-    /**
-     *
-     */
-    public void testCacheContention() throws Exception {
-        int cnt = 10;
-
-        final ExecutorService svc = Executors.newFixedThreadPool(cnt);
-
-        try {
-            Ignite ignite = startGrids(2);
-
-            ignite.cluster().active(true);
-
-            final IgniteCache<Object, Object> cache = ignite.createCache(new CacheConfiguration<>()
-                .setAffinity(new RendezvousAffinityFunction(false, 32))
-                .setAtomicityMode(TRANSACTIONAL)
-                .setBackups(1)
-                .setName(DEFAULT_CACHE_NAME));
-
-            final CountDownLatch l = new CountDownLatch(1);
-
-            final CountDownLatch l2 = new CountDownLatch(1);
-
-            svc.submit(new Runnable() {
-                @Override public void run() {
-                    try (final Transaction tx = ignite.transactions().txStart()) {
-                        cache.put(0, 0);
-
-                        l.countDown();
-
-                        U.awaitQuiet(l2);
-
-                        tx.commit();
-                    }
-                }
-            });
-
-            for (int i = 0; i < cnt - 1; i++) {
-                svc.submit(new Runnable() {
-                    @Override public void run() {
-                        U.awaitQuiet(l);
-
-                        try (final Transaction tx = ignite.transactions().txStart()) {
-                            cache.get(0);
-
-                            tx.commit();
-                        }
-                    }
-                });
-            }
-
-            U.awaitQuiet(l);
-
-            Thread.sleep(300);
-
-            injectTestSystemOut();
-
-            assertEquals(EXIT_CODE_OK, execute("--cache", "contention", "5"));
-
-            l2.countDown();
-
-            String out = testOut.toString();
-
-            assertContains(log, out, "TxEntry");
-            assertContains(log, out, "op=READ");
-            assertContains(log, out, "op=CREATE");
-            assertContains(log, out, "id=" + ignite(0).cluster().localNode().id());
-            assertContains(log, out, "id=" + ignite(1).cluster().localNode().id());
-        }
-        finally {
-            svc.shutdown();
-            svc.awaitTermination(100, TimeUnit.DAYS);
-        }
-    }
-
-    /**
-     *
-     */
+    /** */
+    @Test
     public void testCacheSequence() throws Exception {
         Ignite ignite = startGrid();
 
@@ -2229,226 +1176,6 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
 
         assertContains(log, out, "testSeq");
         assertContains(log, out, "testSeq2");
-    }
-
-    /**
-     *
-     */
-    public void testCacheGroups() throws Exception {
-        Ignite ignite = startGrid();
-
-        ignite.cluster().active(true);
-
-        IgniteCache<Object, Object> cache = ignite.createCache(new CacheConfiguration<>()
-            .setAffinity(new RendezvousAffinityFunction(false, 32))
-            .setBackups(1)
-            .setGroupName("G100")
-            .setName(DEFAULT_CACHE_NAME));
-
-        for (int i = 0; i < 100; i++)
-            cache.put(i, i);
-
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "list", ".*", "--groups"));
-
-        assertContains(log, testOut.toString(), "G100");
-    }
-
-    /**
-     *
-     */
-    public void testCacheAffinity() throws Exception {
-        Ignite ignite = startGrid();
-
-        ignite.cluster().active(true);
-
-        IgniteCache<Object, Object> cache1 = ignite.createCache(new CacheConfiguration<>()
-            .setAffinity(new RendezvousAffinityFunction(false, 32))
-            .setBackups(1)
-            .setName(DEFAULT_CACHE_NAME));
-
-        for (int i = 0; i < 100; i++)
-            cache1.put(i, i);
-
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "list", ".*"));
-
-        String out = testOut.toString();
-
-        assertContains(log, out, "cacheName=" + DEFAULT_CACHE_NAME);
-        assertContains(log, out, "prim=32");
-        assertContains(log, out, "mapped=32");
-        assertContains(log, out, "affCls=RendezvousAffinityFunction");
-    }
-
-    /** */
-    public void testCacheConfigNoOutputFormat() throws Exception {
-        testCacheConfig(null, 1, 1);
-    }
-
-    /** */
-    public void testCacheConfigSingleLineOutputFormatSingleNodeSignleCache() throws Exception {
-        testCacheConfigSingleLineOutputFormat(1, 1);
-    }
-
-    /** */
-    public void testCacheConfigSingleLineOutputFormatTwoNodeSignleCache() throws Exception {
-        testCacheConfigSingleLineOutputFormat(2, 1);
-    }
-
-    /** */
-    public void testCacheConfigSingleLineOutputFormatTwoNodeManyCaches() throws Exception {
-        testCacheConfigSingleLineOutputFormat(2, 100);
-    }
-
-    /** */
-    public void testCacheConfigMultiLineOutputFormatSingleNodeSingleCache() throws Exception {
-        testCacheConfigMultiLineOutputFormat(1, 1);
-    }
-
-    /** */
-    public void testCacheConfigMultiLineOutputFormatTwoNodeSingleCache() throws Exception {
-        testCacheConfigMultiLineOutputFormat(2, 1);
-    }
-
-    /** */
-    public void testCacheConfigMultiLineOutputFormatTwoNodeManyCaches() throws Exception {
-        testCacheConfigMultiLineOutputFormat(2, 100);
-    }
-
-    /** */
-    private void testCacheConfigSingleLineOutputFormat(int nodesCnt, int cachesCnt) throws Exception {
-        testCacheConfig("single-line", nodesCnt, cachesCnt);
-    }
-
-    /** */
-    private void testCacheConfigMultiLineOutputFormat(int nodesCnt, int cachesCnt) throws Exception {
-        testCacheConfig("multi-line", nodesCnt, cachesCnt);
-    }
-
-    /** */
-    private void testCacheConfig(String outputFormat, int nodesCnt, int cachesCnt) throws Exception {
-        assertTrue("Invalid number of nodes or caches", nodesCnt > 0 && cachesCnt > 0);
-
-        Ignite ignite = startGrid(nodesCnt);
-
-        ignite.cluster().active(true);
-
-        List<CacheConfiguration> ccfgs = new ArrayList<>(cachesCnt);
-
-        for (int i = 0; i < cachesCnt; i++) {
-            ccfgs.add(
-                new CacheConfiguration<>()
-                    .setAffinity(new RendezvousAffinityFunction(false, 32))
-                    .setBackups(1)
-                    .setName(DEFAULT_CACHE_NAME + i)
-            );
-        }
-
-        ignite.createCaches(ccfgs);
-
-        IgniteCache<Object, Object> cache1 = ignite.cache(DEFAULT_CACHE_NAME + 0);
-
-        for (int i = 0; i < 100; i++)
-            cache1.put(i, i);
-
-        injectTestSystemOut();
-
-        int exitCode;
-
-        if (outputFormat == null)
-            exitCode = execute("--cache", "list", ".*", "--config");
-        else
-            exitCode = execute("--cache", "list", ".*", "--config", "--output-format", outputFormat);
-
-        assertEquals(EXIT_CODE_OK, exitCode);
-
-        String outStr = testOut.toString();
-
-        if (outputFormat == null || SINGLE_LINE.text().equals(outputFormat)) {
-            for (int i = 0; i < cachesCnt; i++)
-                assertContains(log, outStr, "name=" + DEFAULT_CACHE_NAME + i);
-
-            assertContains(log, outStr, "partitions=32");
-            assertContains(log, outStr, "function=o.a.i.cache.affinity.rendezvous.RendezvousAffinityFunction");
-        }
-        else if (MULTI_LINE.text().equals(outputFormat)) {
-            for (int i = 0; i < cachesCnt; i++)
-                assertContains(log, outStr, "[cache = '" + DEFAULT_CACHE_NAME + i + "']");
-
-            assertContains(log, outStr, "Affinity Partitions: 32");
-            assertContains(log, outStr, "Affinity Function: o.a.i.cache.affinity.rendezvous.RendezvousAffinityFunction");
-        }
-        else
-            fail("Unknown output format: " + outputFormat);
-    }
-
-    /**
-     *
-     */
-    public void testCacheDistribution() throws Exception {
-        Ignite ignite = startGrids(2);
-
-        ignite.cluster().active(true);
-
-        createCacheAndPreload(ignite, 100);
-
-        injectTestSystemOut();
-
-        // Run distribution for all node and all cache
-        assertEquals(EXIT_CODE_OK, execute("--cache", "distribution", "null"));
-
-        String out = testOut.toString();
-
-        // Result include info by cache "default"
-        assertContains(log ,out, "[next group: id=1544803905, name=default]");
-
-        // Result include info by cache "ignite-sys-cache"
-        assertContains(log ,out, "[next group: id=-2100569601, name=ignite-sys-cache]");
-
-        // Run distribution for all node and all cache and include additional user attribute
-        assertEquals(EXIT_CODE_OK, execute("--cache", "distribution", "null", "--user-attributes", "ZONE,CELL,DC"));
-
-        out = testOut.toString();
-
-        List<String> outLines = Arrays.stream(out.split("\n"))
-                                .map(String::trim)
-                                .collect(Collectors.toList());
-
-        int firstIndex = outLines.indexOf("[next group: id=1544803905, name=default]");
-        int lastIndex  = outLines.lastIndexOf("[next group: id=1544803905, name=default]");
-
-        String dataLine = outLines.get(firstIndex + 1);
-        String userArrtDataLine = outLines.get(lastIndex + 1);
-
-        long commaNum = dataLine.chars().filter(i -> i == ',').count();
-        long userArrtCommaNum = userArrtDataLine.chars().filter(i -> i == ',').count();
-
-        // Check that number of columns increased by 3
-        assertEquals(3, userArrtCommaNum - commaNum);
-    }
-
-    /**
-     *
-     */
-    public void testCacheResetLostPartitions() throws Exception {
-        Ignite ignite = startGrids(2);
-
-        ignite.cluster().active(true);
-
-        createCacheAndPreload(ignite, 100);
-
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--cache", "reset_lost_partitions", "ignite-sys-cache,default"));
-
-        final String out = testOut.toString();
-
-        assertContains(log, out, "Reset LOST-partitions performed successfully. Cache group (name = 'ignite-sys-cache'");
-
-        assertContains(log, out, "Reset LOST-partitions performed successfully. Cache group (name = 'default'");
     }
 
     /**
@@ -2477,84 +1204,11 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
     }
 
     /**
-     * Test execution of --wal print command.
-     *
-     * @throws Exception if failed.
-     */
-    public void testUnusedWalPrint() throws Exception {
-        Ignite ignite = startGrids(2);
-
-        ignite.cluster().active(true);
-
-        List<String> nodes = new ArrayList<>(2);
-
-        for (ClusterNode node : ignite.cluster().forServers().nodes())
-            nodes.add(node.consistentId().toString());
-
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--wal", "print"));
-
-        String out = testOut.toString();
-
-        for (String id : nodes)
-            assertContains(log, out, id);
-
-        assertNotContains(log, out, "error");
-
-        testOut.reset();
-
-        assertEquals(EXIT_CODE_OK, execute("--wal", "print", nodes.get(0)));
-
-        out = testOut.toString();
-
-        assertNotContains(log, out, nodes.get(1));
-
-        assertNotContains(log, out, "error");
-    }
-
-    /**
-     * Test execution of --wal delete command.
-     *
-     * @throws Exception if failed.
-     */
-    public void testUnusedWalDelete() throws Exception {
-        Ignite ignite = startGrids(2);
-
-        ignite.cluster().active(true);
-
-        List<String> nodes = new ArrayList<>(2);
-
-        for (ClusterNode node : ignite.cluster().forServers().nodes())
-            nodes.add(node.consistentId().toString());
-
-        injectTestSystemOut();
-
-        assertEquals(EXIT_CODE_OK, execute("--wal", "delete"));
-
-        String out = testOut.toString();
-
-        for (String id : nodes)
-            assertContains(log, out, id);
-
-        assertNotContains(log, out, "error");
-
-        testOut.reset();
-
-        assertEquals(EXIT_CODE_OK, execute("--wal", "delete", nodes.get(0)));
-
-        out = testOut.toString();
-
-        assertNotContains(log, out, nodes.get(1));
-
-        assertNotContains(log, out, "error");
-    }
-
-    /**
      * Test execution of --diagnostic command.
      *
      * @throws Exception if failed.
      */
+    @Test
     public void testDiagnosticPageLocksTracker() throws Exception {
         Ignite ignite = startGrids(4);
 
@@ -2826,38 +1480,72 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
         }
     }
 
-    /**
-     * Don't show wal commands by --help in case
-     * {@link org.apache.ignite.IgniteSystemProperties#IGNITE_ENABLE_EXPERIMENTAL_COMMAND} = false or empty.
-     */
+    /** */
     @Test
-    public void testHideWalInHelpWhenDisableExperimentalCommand() {
-        System.clearProperty(IGNITE_ENABLE_EXPERIMENTAL_COMMAND);
+    public void testKillHangingLocalTransactions() throws Exception {
+        Ignite ignite = startGridsMultiThreaded(2);
 
-        injectTestSystemOut();
+        ignite.cluster().active(true);
 
-        execute("--help");
+        Ignite client = startGrid("client");
 
-        assertNotContains(log, testOut.toString(), WAL.text());
-    }
+        client.getOrCreateCache(new CacheConfiguration<>(DEFAULT_CACHE_NAME).
+            setAtomicityMode(TRANSACTIONAL).
+            setWriteSynchronizationMode(FULL_SYNC).
+            setAffinity(new RendezvousAffinityFunction(false, 64)));
 
-    /**
-     * Wal commands should ignored and print warning in case
-     * {@link org.apache.ignite.IgniteSystemProperties#IGNITE_ENABLE_EXPERIMENTAL_COMMAND} = false or empty.
-     * */
-    @Test
-    public void testWalCommandsInCaseDisableExperimentalCommand() {
-        System.clearProperty(IGNITE_ENABLE_EXPERIMENTAL_COMMAND);
+        Ignite prim = primaryNode(0L, DEFAULT_CACHE_NAME);
 
-        injectTestSystemOut();
+        // Blocks lock response to near node.
+        TestRecordingCommunicationSpi.spi(prim).blockMessages(GridNearLockResponse.class, client.name());
 
-        String warning = String.format("For use experimental command add %s=true to JVM_OPTS in %s",
-            IGNITE_ENABLE_EXPERIMENTAL_COMMAND, UTILITY_NAME);
+        TestRecordingCommunicationSpi.spi(client).blockMessages(GridNearTxFinishRequest.class, prim.name());
 
-        Stream.of("print", "delete")
-            .peek(c -> testOut.reset())
-            .peek(c -> assertEquals(EXIT_CODE_OK, execute(WAL.text(), c)))
-            .forEach(c -> assertContains(log, testOut.toString(), warning));
+        GridNearTxLocal clientTx = null;
+
+        try (Transaction tx = client.transactions().txStart(PESSIMISTIC, READ_COMMITTED, 2000, 1)) {
+            clientTx = ((TransactionProxyImpl)tx).tx();
+
+            client.cache(DEFAULT_CACHE_NAME).put(0L, 0L);
+
+            fail();
+        }
+        catch (Exception e) {
+            assertTrue(X.hasCause(e, TransactionTimeoutException.class));
+        }
+
+        assertNotNull(clientTx);
+
+        IgniteEx primEx = (IgniteEx)prim;
+
+        IgniteInternalTx tx0 = primEx.context().cache().context().tm().activeTransactions().iterator().next();
+
+        assertNotNull(tx0);
+
+        CommandHandler h = new CommandHandler();
+
+        validate(h, map -> {
+            ClusterNode node = grid(0).cluster().localNode();
+
+            VisorTxTaskResult res = map.get(node);
+
+            for (VisorTxInfo info : res.getInfos())
+                assertEquals(tx0.xid(), info.getXid());
+
+            assertEquals(1, map.size());
+        }, "--tx", "--xid", tx0.xid().toString(), "--kill");
+
+        tx0.finishFuture().get();
+
+        TestRecordingCommunicationSpi.spi(prim).stopBlock();
+
+        TestRecordingCommunicationSpi.spi(client).stopBlock();
+
+        IgniteInternalFuture<?> nearFinFut = U.field(clientTx, "finishFut");
+
+        nearFinFut.get();
+
+        checkUserFutures();
     }
 
     /**
@@ -2878,5 +1566,44 @@ public class GridCommandHandlerTest extends GridCommandHandlerAbstractTest {
 
         assertEquals(EXIT_CODE_UNEXPECTED_ERROR,
             execute("--baseline", "set", "non-existing-node-id ," + consistentIds(ignite)));
+    }
+
+    /**
+     * @throws Exception If failed.
+     */
+    @Test
+    public void testCacheIdleVerifyPrintLostPartitions() throws Exception {
+        IgniteEx ignite = startGrids(3);
+
+        ignite.cluster().active(true);
+
+        ignite.createCache(new CacheConfiguration<>(DEFAULT_CACHE_NAME)
+            .setAffinity(new RendezvousAffinityFunction(false, 16))
+            .setCacheMode(PARTITIONED)
+            .setPartitionLossPolicy(READ_ONLY_SAFE)
+            .setBackups(1));
+
+        try (IgniteDataStreamer streamer = ignite.dataStreamer(DEFAULT_CACHE_NAME)) {
+            for (int i = 0; i < 10000; i++)
+                streamer.addData(i, new byte[i]);
+        }
+
+        String g1Name = grid(1).name();
+
+        stopGrid(1);
+
+        cleanPersistenceDir(g1Name);
+
+        //Start node 2 with empty PDS. Rebalance will be started.
+        startGrid(1);
+
+        //During rebalance stop node 3. Rebalance will be stopped which lead to lost partitions.
+        stopGrid(2);
+
+        injectTestSystemOut();
+
+        assertEquals(EXIT_CODE_OK, execute("--cache", "idle_verify", "--yes"));
+
+        assertContains(log, testOut.toString(), "LOST partitions:");
     }
 }
