+++ modules/indexing/src/test/java/org/apache/ignite/testsuites/IgniteCacheWithIndexingAndPersistenceTestSuite.java
@@ -20,6 +20,7 @@ package org.apache.ignite.testsuites;
 import junit.framework.TestSuite;
 import org.apache.ignite.internal.processors.cache.StartCachesInParallelTest;
 import org.apache.ignite.internal.processors.cache.index.IoStatisticsBasicIndexSelfTest;
+import org.apache.ignite.util.GridCommandHandlerIndexingClusterByClassTest;
 import org.apache.ignite.util.GridCommandHandlerIndexingTest;
 
 /**
@@ -34,6 +35,7 @@ public class IgniteCacheWithIndexingAndPersistenceTestSuite extends TestSuite {
         TestSuite suite = new TestSuite("Ignite Cache With Indexing And Persistence Test Suite");
 
         suite.addTestSuite(GridCommandHandlerIndexingTest.class);
+        suite.addTestSuite(GridCommandHandlerIndexingClusterByClassTest.class);
         suite.addTestSuite(StartCachesInParallelTest.class);
         suite.addTestSuite(IoStatisticsBasicIndexSelfTest.class);
